Add feature_extraction
- Add HashingVectorizer
- Add support for dask.bag.Bag as inputs
